NWT-51 Update DB schema and code references

+++ app/src/main/java/com/meadowlandapps/simplelists/ui/AddTabFragment.kt
@@ -19,12 +19,12 @@ import com.meadowlandapps.simplelists.R
 
 class AddTabFragment : Fragment() {
 
-    private var categoryId: Int = 0
+    private var categoryId: Long = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            categoryId = it.getInt(BUNDLE_KEY_CATEGORY, 0)
+            categoryId = it.getLong(BUNDLE_KEY_CATEGORY, 0)
         }
     }
