+++ app/src/main/java/com/meadowlandapps/simplelists/ui/DeleteItemsConfirmDialogFragment.kt
@@ -12,12 +12,12 @@ import com.meadowlandapps.simplelists.R
  * A confirmation dialog when deleting items
  */
 class DeleteItemsConfirmDialogFragment : DialogFragment() {
-    private lateinit var items: IntArray
+    private lateinit var items: LongArray
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let { bundle ->
-            items = bundle.getIntArray(BUNDLE_KEY_ITEMS) ?: IntArray(0)
+            items = bundle.getLongArray(BUNDLE_KEY_ITEMS) ?: LongArray(0)
         }
     }
